+++ src/demo/ts/ephox/alloy/demo/Demos.ts
@@ -18,6 +18,7 @@ import InspectorDemo from './InspectorDemo';
 import LongpressDemo from './LongpressDemo';
 import PinchingDemo from './PinchingDemo';
 import PositionDemo from './PositionDemo';
+import ReflectingDemo from './ReflectingDemo';
 import SliderDemo from './SliderDemo';
 import SlidingDemo from './SlidingDemo';
 import TabSectionDemo from './TabSectionDemo';
@@ -49,6 +50,7 @@ window.demos = {
   LongpressDemo,
   PinchingDemo,
   PositionDemo,
+  ReflectingDemo,
   SliderDemo,
   SlidingDemo,
   SlotContainerDemo,